file reorg, nuget updates
