+++ Framework/Test/FakeTaskActivityExecutor.cs
@@ -16,6 +16,8 @@ namespace DurableTask.Test
     using System;
     using System.Threading;
     using System.Threading.Tasks;
+    using Common;
+    using Exceptions;
 
     internal class FakeTaskActivityExecutor
     {
