+++ Framework/Test/OrchestrationTestHost.cs
@@ -14,10 +14,10 @@
 namespace DurableTask.Test
 {
     using System;
-    using System.Diagnostics;
     using System.Reflection;
     using System.Threading;
     using System.Threading.Tasks;
+    using Exceptions;
 
     public class OrchestrationTestHost
     {
