+++ Framework/Tracking/OrchestrationStateEntity.cs
@@ -15,6 +15,7 @@ namespace DurableTask.Tracking
 {
     using System;
     using System.Collections.Generic;
+    using Common;
     using DurableTask.History;
     using Microsoft.WindowsAzure.Storage;
     using Microsoft.WindowsAzure.Storage.Table;
