+++ FrameworkUnitTests/OrchestrationTestHostTests.cs
@@ -20,6 +20,7 @@ namespace FrameworkUnitTests
     using System.Linq;
 
     using DurableTask;
+    using DurableTask.Exceptions;
     using DurableTask.Test;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
