+++ FrameworkUnitTests/packages.config
@@ -5,8 +5,8 @@
   <package id="Microsoft.Data.OData" version="5.2.0" targetFramework="net45" />
   <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net45" />
   <package id="Microsoft.WindowsAzure.ConfigurationManager" version="2.0.3" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="5.0.6" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net45" />
   <package id="System.Spatial" version="5.2.0" targetFramework="net45" />
-  <package id="WindowsAzure.ServiceBus" version="3.0.1" targetFramework="net45" />
+  <package id="WindowsAzure.ServiceBus" version="3.0.9" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="2.1.0.4" targetFramework="net45" />
 </packages>
